@@ -26,7 +26,7 @@ import re
 import sys
 from contextlib import suppress, contextmanager
 from types import ModuleType as ModelModule
-from typing import Optional, Union, Sequence, List
+from typing import Any, Dict, Optional, Union, Sequence, List
 
 
 def _get_amici_path():
@@ -127,13 +127,7 @@ if not _imported_from_setup():
 
         # These module require the swig interface and other dependencies
         from .numpy import ReturnDataView, ExpDataView
-        from .pandas import (
-            getEdataFromDataFrame,
-            getDataObservablesAsDataFrame,
-            getSimulationObservablesAsDataFrame,
-            getSimulationStatesAsDataFrame,
-            getResidualsAsDataFrame
-        )
+        from .pandas import *
 
     # These modules don't require the swig interface
     from .sbml_import import SbmlImporter, assignmentRules2observables
@@ -144,7 +138,8 @@ if not _imported_from_setup():
         from typing import Protocol
 
         class ModelModule(Protocol):
-            """Enable Python static type checking for AMICI-generated model modules"""
+            """Enable Python static type checking for AMICI-generated model
+            modules"""
             def getModel(self) -> amici.Model:
                 pass
     except ImportError:
@@ -278,6 +273,65 @@ def writeSolverSettingsToHDF5(
     :param location: location of solver settings in hdf5 file
     """
     amici.writeSolverSettingsToHDF5(_get_ptr(solver), file, location)
+
+
+# Values are suffixes of `get[...]` and `set[...]` `amici.Model` methods.
+# If either the getter or setter is not named with this pattern, then the value
+# is a tuple where the first and second elements are the getter and setter
+# methods, respectively.
+model_instance_settings = [
+    'AddSigmaResiduals',
+    'AlwaysCheckFinite',
+    'FixedParameters',
+    'InitialStates',
+    'InitialStateSensitivities',
+    'MinimumSigmaResiduals',
+    ('nMaxEvent', 'setNMaxEvent'),
+    'Parameters',
+    'ParameterList',
+    'ParameterScale',  # getter returns a SWIG object
+    'ReinitializationStateIdxs',
+    'ReinitializeFixedParameterInitialStates',
+    'StateIsNonNegative',
+    'SteadyStateSensitivityMode',
+    ('t0', 'setT0'),
+    'Timepoints',
+]
+
+
+def get_model_settings(
+        model: AmiciModel,
+) -> Dict[str, Any]:
+    """Get model settings that are set independently of the compiled model.
+
+    :param model: The AMICI model instance.
+
+    :returns: Keys are AMICI model attributes, values are attribute values.
+    """
+    settings = {}
+    for setting in model_instance_settings:
+        getter = setting[0] if isinstance(setting, tuple) else f'get{setting}'
+        settings[setting] = getattr(model, getter)()
+        # TODO `amici.Model.getParameterScale` returns a SWIG object instead
+        # of a Python list/tuple.
+        if setting == 'ParameterScale':
+            settings[setting] = tuple(settings[setting])
+    return settings
+
+
+def set_model_settings(
+        model: AmiciModel,
+        settings: Dict[str, Any],
+) -> None:
+    """Set model settings.
+
+    :param model: The AMICI model instance.
+    :param settings: Keys are callable attributes (setters) of an AMICI model,
+        values are provided to the setters.
+    """
+    for setting, value in settings.items():
+        setter = setting[1] if isinstance(setting, tuple) else f'set{setting}'
+        getattr(model, setter)(value)
 
 
 class add_path:
