@@ -14,20 +14,17 @@ import copy
 from typing import Callable, Optional, List, Sequence
 
 
-def assert_fun(x):
-    assert x
-
-
-def check_finite_difference(x0: Sequence[float],
-                            model: Model,
-                            solver: Solver,
-                            edata: ExpData,
-                            ip: int,
-                            fields: List[str],
-                            assert_fun: Callable,
-                            atol: Optional[float] = 1e-4,
-                            rtol: Optional[float] = 1e-4,
-                            epsilon: Optional[float] = 1e-3) -> None:
+def check_finite_difference(
+        x0: Sequence[float],
+        model: Model,
+        solver: Solver,
+        edata: ExpData,
+        ip: int,
+        fields: List[str],
+        atol: Optional[float] = 1e-4,
+        rtol: Optional[float] = 1e-4,
+        epsilon: Optional[float] = 1e-3
+) -> None:
     """
     Checks the computed sensitivity based derivatives against a finite
     difference approximation.
@@ -49,10 +46,6 @@ def check_finite_difference(x0: Sequence[float],
 
     :param fields:
         rdata fields for which to check the gradient
-
-    :param assert_fun:
-        function that asserts the return values of comparison, enables
-        passing of custom assert function from testing frameworks
 
     :param atol:
         absolute tolerance for comparison
@@ -83,7 +76,8 @@ def check_finite_difference(x0: Sequence[float],
     if int(og_sensitivity_order) < int(SensitivityOrder.first):
         solver.setSensitivityOrder(SensitivityOrder.first)
     rdata = runAmiciSimulation(model, solver, edata)
-    assert_fun(rdata['status'] == AMICI_SUCCESS)
+    if rdata['status'] != AMICI_SUCCESS:
+        raise AssertionError(f"Simulation failed (status {rdata['status']}")
 
     # finite difference
     solver.setSensitivityOrder(SensitivityOrder.none)
@@ -101,16 +95,18 @@ def check_finite_difference(x0: Sequence[float],
     # forward:
     model.setParameters(pf)
     rdataf = runAmiciSimulation(model, solver, edata)
-    assert_fun(rdataf['status'] == AMICI_SUCCESS)
+    if rdataf['status'] != AMICI_SUCCESS:
+        raise AssertionError(f"Simulation failed (status {rdataf['status']}")
 
     # backward:
     model.setParameters(pb)
     rdatab = runAmiciSimulation(model, solver, edata)
-    assert_fun(rdatab['status'] == AMICI_SUCCESS)
+    if rdatab['status'] != AMICI_SUCCESS:
+        raise AssertionError(f"Simulation failed (status {rdatab['status']}")
 
     for field in fields:
         sensi_raw = rdata[f's{field}']
-        fd = (rdataf[field]-rdatab[field])/(pf[ip] - pb[ip])
+        fd = (rdataf[field] - rdatab[field]) / (pf[ip] - pb[ip])
         if len(sensi_raw.shape) == 1:
             sensi = sensi_raw[0]
         elif len(sensi_raw.shape) == 2:
@@ -118,10 +114,9 @@ def check_finite_difference(x0: Sequence[float],
         elif len(sensi_raw.shape) == 3:
             sensi = sensi_raw[:, 0, :]
         else:
-            assert_fun(False)  # not implemented
-            return
+            raise NotImplementedError()
 
-        check_close(sensi, fd, assert_fun, atol, rtol, field, ip=ip)
+        _check_close(sensi, fd, atol=atol, rtol=rtol, field=field, ip=ip)
 
     solver.setSensitivityOrder(og_sensitivity_order)
     model.setParameters(og_parameters)
@@ -130,15 +125,16 @@ def check_finite_difference(x0: Sequence[float],
         edata.plist = og_eplist
 
 
-def check_derivatives(model: Model,
-                      solver: Solver,
-                      edata: Optional[ExpData] = None,
-                      assert_fun: Optional[Callable] = assert_fun,
-                      atol: Optional[float] = 1e-4,
-                      rtol: Optional[float] = 1e-4,
-                      epsilon: Optional[float] = 1e-3,
-                      check_least_squares: bool = True,
-                      skip_zero_pars: bool = False) -> None:
+def check_derivatives(
+        model: Model,
+        solver: Solver,
+        edata: Optional[ExpData] = None,
+        atol: Optional[float] = 1e-4,
+        rtol: Optional[float] = 1e-4,
+        epsilon: Optional[float] = 1e-3,
+        check_least_squares: bool = True,
+        skip_zero_pars: bool = False
+) -> None:
     """
     Finite differences check for likelihood gradient.
 
@@ -150,10 +146,6 @@ def check_derivatives(model: Model,
 
     :param edata:
         exp data
-
-    :param assert_fun:
-        function that asserts the return values of comparison, enables
-        passing of custom assert function from testing frameworks
 
     :param atol:
         absolute tolerance for comparison
@@ -180,11 +172,10 @@ def check_derivatives(model: Model,
     rdata = runAmiciSimulation(model, solver, edata)
     solver.setSensitivityOrder(og_sens_order)
 
-    assert_fun(rdata['status'] == AMICI_SUCCESS)
+    if rdata['status'] != AMICI_SUCCESS:
+        raise AssertionError(f"Simulation failed (status {rdata['status']}")
 
     fields = []
-    if edata is not None:
-        fields.append('llh')
 
     if solver.getSensitivityMethod() == SensitivityMethod.forward and \
             solver.getSensitivityOrder() <= SensitivityOrder.first:
@@ -202,29 +193,30 @@ def check_derivatives(model: Model,
     if check_least_squares and leastsquares_applicable:
         fields += ['res', 'y']
 
-        check_results(rdata, 'FIM',
-                      np.dot(rdata['sres'].T, rdata['sres']),
-                      assert_fun,
-                      1e-8, 1e-4)
-        check_results(rdata, 'sllh',
-                      -np.dot(rdata['res'].T, rdata['sres']),
-                      assert_fun,
-                      1e-8, 1e-4)
+        _check_results(rdata, 'FIM', np.dot(rdata['sres'].T, rdata['sres']),
+                       atol=1e-8, rtol=1e-4)
+        _check_results(rdata, 'sllh', -np.dot(rdata['res'].T, rdata['sres']),
+                       atol=1e-8, rtol=1e-4)
+
+    if edata is not None:
+        fields.append('llh')
+
     for ip, pval in enumerate(p):
         if pval == 0.0 and skip_zero_pars:
             continue
         check_finite_difference(p, model, solver, edata, ip, fields,
-                                assert_fun, atol=atol, rtol=rtol,
-                                epsilon=epsilon)
+                                atol=atol, rtol=rtol, epsilon=epsilon)
 
 
-def check_close(result: np.array,
-                expected: np.array,
-                assert_fun: Callable,
-                atol: float,
-                rtol: float,
-                field: str,
-                ip: Optional[int] = None) -> None:
+def _check_close(
+        result: np.array,
+        expected: np.array,
+        atol: float,
+        rtol: float,
+        field: str,
+        ip: Optional[int] = None,
+        verbose: Optional[bool] = True,
+) -> None:
     """
     Compares computed values against expected values and provides rich
     output information.
@@ -238,10 +230,6 @@ def check_close(result: np.array,
     :param field:
         rdata field for which the gradient is checked, only for error reporting
 
-    :param assert_fun:
-        function that asserts the return values of comparison, enables
-        passing of custom assert function from testing frameworks
-
     :param atol:
         absolute tolerance for comparison
 
@@ -249,33 +237,43 @@ def check_close(result: np.array,
         relative tolerance for comparison
 
     :param ip:
-        parameter index
+        parameter index, for more informative output
 
+    :param verbose:
+        produce a more verbose error message in case of unmatched expectations
     """
     close = np.isclose(result, expected, atol=atol, rtol=rtol, equal_nan=True)
+    if close.all():
+        return
 
-    if not close.all():
-        if ip is None:
-            index_str = ''
-            check_type = 'Regression check  '
-        else:
-            index_str = f'at index ip={ip} '
-            check_type = 'FD check '
-        print(f'{check_type} failed for {field} {index_str}for '
-              f'{close.size - close.sum()} indices:')
-        adev = np.abs(result - expected)
-        rdev = np.abs((result - expected)/(expected + atol))
-        print(f'max(adev): {adev.max()}, max(rdev): {rdev.max()}')
+    if ip is None:
+        index_str = ''
+        check_type = 'Regression check'
+    else:
+        index_str = f'at index ip={ip} '
+        check_type = 'FD check'
 
-    assert_fun(close.all())
+    lines = [f'{check_type} failed for {field} {index_str}for '
+             f'{close.size - close.sum()} indices:']
+    if verbose:
+        for idx in np.argwhere(~close):
+            idx = tuple(idx)
+            lines.append(
+                f"\tat {idx}: Expected {expected[idx]}, got {result[idx]}")
+    adev = np.abs(result - expected)
+    rdev = np.abs((result - expected) / (expected + atol))
+    lines.append(f'max(adev): {adev.max()}, max(rdev): {rdev.max()}')
+
+    raise AssertionError("\n".join(lines))
 
 
-def check_results(rdata: ReturnData,
-                  field: str,
-                  expected: np.array,
-                  assert_fun: Callable,
-                  atol: float,
-                  rtol: float) -> None:
+def _check_results(
+        rdata: ReturnData,
+        field: str,
+        expected: np.array,
+        atol: float,
+        rtol: float
+        ) -> None:
     """
     Checks whether rdata[field] agrees with expected according to provided
     tolerances.
@@ -290,10 +288,6 @@ def check_results(rdata: ReturnData,
     :param expected:
         expected values
 
-    :param assert_fun:
-        function that asserts the return values of comparison, enables
-        passing of custom assert function from testing frameworks
-
     :param atol:
         absolute tolerance for comparison
 
@@ -305,4 +299,5 @@ def check_results(rdata: ReturnData,
     if type(result) is float:
         result = np.array(result)
 
-    check_close(result, expected, assert_fun, atol, rtol, field)
+    _check_close(result=result, expected=expected,
+                 atol=atol, rtol=rtol, field=field)
