@@ -2,7 +2,10 @@
 PEtab Simulate
 --------------
 Functionality related to the use of AMICI for simulation with PEtab's
-Simulator class. Use cases:
+Simulator class.
+
+Use cases:
+
 - generate data for use with PEtab's plotting methods
 - generate synthetic data
 """
