@@ -25,40 +25,12 @@
 
 namespace amici {
 
-/**
- * c++ interface to the isNaN function
- *
- * @param what argument
- * @return isnan(what)
- *
- */
 int isNaN(double what) { return std::isnan(what); }
 
-/**
- * c++ interface to the isinf function
- *
- * @param what argument
- * @return isnan(what)
- *
- */
 int isInf(double what) { return std::isinf(what); }
 
-/**
- * function returning nan
- *
- * @return NaN
- *
- */
 double getNaN() { return NAN; }
 
-/**
- * C implementation of log function, this prevents returning NaN values for
- * negative values
- *
- * @param x argument
- * @return if(x>0) then log(x) else -Inf
- *
- */
 double log(double x) {
     if (x <= 0) {
         return -std::log(DBL_MAX);
@@ -66,13 +38,6 @@ double log(double x) {
     return std::log(x);
 }
 
-/**
- * C implementation of matlab function dirac
- *
- * @param x argument
- * @return if(x==0) then INF else 0
- *
- */
 double dirac(double x) {
     if (x == 0.0) {
         return DBL_MAX;
@@ -84,23 +49,18 @@ double dirac(double x) {
  * c implementation of matlab function heaviside
  *
  * @param x argument
- * @return if(x>0) then 1 else 0
+ * @param x0 value at x==0
+ * @return if(x>0) then 1 else if (x==0) then x0 else 0
  *
  */
-double heaviside(double x) {
-    if (x < 0.0) {
+double heaviside(double x, double x0) {
+    if (x < 0.0)
         return 0.0;
-    }
+    if (x == 0.0)
+        return x0;
     return 1.0;
 }
 
-/**
- *  c implementation of matlab function sign
- *
- * @param x argument
- * @return 0
- *
- */
 double sign(double x) {
     if (x > 0.0)
         return 1.0;
@@ -111,14 +71,6 @@ double sign(double x) {
     return 0.0;
 }
 
-/**
- * c implementation of matlab function max
- *
- * @param a value1
- * @param b value2
- * @return if(a > b) then a else b
- *
- */
 double max(double a, double b, double  /*c*/) {
     int anan = isNaN(a), bnan = isNaN(b);
     if (anan || bnan) {
@@ -131,29 +83,10 @@ double max(double a, double b, double  /*c*/) {
     return (std::max(a, b));
 }
 
-/**
- * c implementation of matlab function min
- *
- * @param a value1
- * @param b value2
- * @param c bogus parameter to ensure correct parsing as a function
- * @return if(a < b) then a else b
- *
- */
 double min(double a, double b, double c) {
     return (-max(-a,-b,c));
 }
 
-/**
- * parameter derivative of c implementation of matlab function max
- *
- * @param id argument index for differentiation
- * @param a value1
- * @param b value2
- * @return id == 1:  if(a > b) then 1 else 0
- * @return id == 2:  if(a > b) then 0 else 1
- *
- */
 double Dmax(int id, double a, double b, double  /*c*/) {
     if (id == 1.0) {
         if (a > b)
@@ -166,49 +99,18 @@ double Dmax(int id, double a, double b, double  /*c*/) {
     return 1.0;
 }
 
-/**
- * parameter derivative of c implementation of matlab function max
- *
- * @param id argument index for differentiation
- * @param a value1
- * @param b value2
- * @param c bogus parameter to ensure correct parsing as a function
- * @return id == 1:  if(a > b) then 1 else 0
- * @return id == 2:  if(a > b) then 0 else 1
- *
- */
 double Dmin(int id, double a, double b, double c) {
     return Dmax(id,-a,-b,c);
 }
 
-/**
- * specialized pow functions that assumes positivity of the first argument
- *
- * @param base base
- * @param exponent exponent
- * @return pow(max(base,0.0),exponent)
- *
- */
 double pos_pow(double base, double exponent) {
     // we do NOT want to propagate NaN values here, if base is nan, so should the output be
     return pow(std::max(base, 0.0),exponent);
 }
 
-/**
- * @brief Spline function
- *
- * Takes variable argument pairs (ti,pi) with `ti`: location of node i and
- * `pi`: spline value at node i. the last two arguments are always `ss`: flag
- * indicating whether slope at first node should be user defined
- * and `dudt` user defined slope at first node. All arguments must be of type
- * double.
- *
- * @param t point at which the spline should be evaluated
- * @param num number of spline nodes
- *
- * @return spline(t)
- *
- */
+
+
+// Legacy spline implementation in C (MATLAB only)
 double spline(double t, int num, ...) {
 
     va_list valist;
@@ -245,21 +147,6 @@ double spline(double t, int num, ...) {
     return uout;
 }
 
-/**
- * @brief Exponentiated spline function
- *
- * Takes variable argument pairs (ti,pi) with `ti`: location of node i and
- * `pi`: spline value at node i. the last two arguments are always `ss`: flag
- * indicating whether slope at first node should be user defined
- * and `dudt` user defined slope at first node. All arguments must be of type
- * double.
- *
- * @param t point at which the spline should be evaluated
- * @param num number of spline nodes
- *
- * @return spline(t)
- *
- */
 double spline_pos(double t, int num, ...) {
 
     va_list valist;
@@ -297,23 +184,6 @@ double spline_pos(double t, int num, ...) {
     return exp(uout);
 }
 
-/**
- * @brief Derivation of a spline function
- *
- * Takes variable argument pairs (ti,pi) with `ti`: location of node i and
- * `pi`: spline value at node i. the last two arguments are always `ss`: flag
- * indicating whether slope at first node should be user defined
- * and `dudt` user defined slope at first node. All arguments but id must be of
- * type double.
- *
- * @param id index of node to which the derivative of the corresponding spline
- * coefficient should be computed
- * @param t point at which the spline should be evaluated
- * @param num number of spline nodes
- *
- * @return dsplinedp(t)
- *
- */
 double Dspline(int id, double t, int num, ...) {
 
     va_list valist;
@@ -354,23 +224,6 @@ double Dspline(int id, double t, int num, ...) {
     return uout;
 }
 
-/**
- * @brief Derivation of an exponentiated spline function
- *
- * Takes variable argument pairs (ti,pi) with `ti`: location of node i and
- * `pi`: spline value at node i. the last two arguments are always `ss`: flag
- * indicating whether slope at first node should be user defined
- * and `dudt` user defined slope at first node. All arguments but id must be of
- * type double.
- *
- * @param id index of node to which the derivative of the corresponding spline
- * coefficient should be computed
- * @param t point at which the spline should be evaluated
- * @param num number of spline nodes
- *
- * @return dsplinedp(t)
- *
- */
 double Dspline_pos(int id, double t, int num, ...) {
 
     va_list valist;
@@ -419,45 +272,8 @@ double Dspline_pos(int id, double t, int num, ...) {
     return uout;
 }
 
-/**
- * @brief Second derivation of a spline function
- *
- * Takes variable argument pairs (ti,pi) with `ti`: location of node i and
- * `pi`: spline value at node i. the last two arguments are always `ss`: flag
- * indicating whether slope at first node should be user defined
- * and `dudt` user defined slope at first node. All arguments but id1 and id2
- * must be of type double.
- *
- * @param id1 index of node to which the first derivative of the corresponding
- * spline coefficient should be computed
- * @param id2 index of node to which the second derivative of the corresponding
- * spline coefficient should be computed
- * @param t point at which the spline should be evaluated
- * @param num number of spline nodes
- *
- * @return ddspline(t)
- */
 double DDspline(int  /*id1*/, int  /*id2*/, double  /*t*/, int  /*num*/, ...) { return 0.0; }
 
-/**
- * @brief Derivation of an exponentiated spline function
- *
- * Takes variable argument pairs (ti,pi) with `ti`: location of node i and
- * `pi`: spline value at node i. the last two arguments are always `ss`: flag
- * indicating whether slope at first node should be user defined
- * and `dudt` user defined slope at first node. All arguments but id1 and id2
- * must be of type double.
- *
- * @param id1 index of node to which the first derivative of the corresponding
- * spline coefficient should be computed
- * @param id2 index of node to which the second derivative of the corresponding
- * spline coefficient should be computed
- * @param t point at which the spline should be evaluated
- * @param num number of spline nodes
- *
- * @return ddspline(t)
- *
- */
 double DDspline_pos(int id1, int id2, double t, int num, ...) {
 
     va_list valist;
